Fixes for getting Mac version to build: include missing prototypes; fix definition of gef_get_c_pointer, fix bug in end-of-file wildcard asterisks; use PyMem_Malloc and PyMem_Free instead of places where malloc and free were still being used
git-svn-id: file:///home/users/tmap/svn/repos/ferret/branches/pyferret@24007 fdbf22ae-c210-0410-be80-ca943da6b8f8

+++ fer/ccr/get_ws_mem.c
@@ -21,7 +21,7 @@
   input  - index:  Ferret ws index at which to store the array pointer
   input  - alen:   array length
 */
-void FORTRAN(get_ws_mem)( int *index, long *alen, int *status )
+void FORTRAN(get_ws_mem)( int *index, int *alen, int *status )
 {
   double *pointer;
